feat: add auth info. to events, move event's helpers to other files
