+++ internal/event/auth.go
@@ -1,0 +1,6 @@
+package event
+
+func (e *Event) WithPassword(hash string) *Event {
+	e.AuthInfo = &AuthInfo{HashedPassword: hash}
+	return e
+}
