+++ internal/event/event_data.go
@@ -1,0 +1,18 @@
+package event
+
+import "encoding/json"
+
+type EventData map[string]any
+
+func (ed *EventData) JSON() ([]byte, error) {
+	rawJSON, err := json.MarshalIndent(ed, "", "  ")
+	if err != nil {
+		return nil, err
+	}
+
+	return rawJSON, nil
+}
+
+func (ed *EventData) Get(key string) any {
+	return (map[string]any)(*ed)[key]
+}
